Load changelog content on demand
